[Ray]Move Patch File
