+++ vllm_ascend/platform.py
@@ -104,7 +104,7 @@ class NPUPlatform(Platform):
     @classmethod
     def check_and_update_config(cls, vllm_config: VllmConfig) -> None:
         # RayWorkerWrapper monkey patch when setup
-        from vllm_ascend import ray_patch  # noqa: F401
+        from vllm_ascend.patch import ray_patch  # noqa: F401
 
         parallel_config = vllm_config.parallel_config
         if parallel_config.worker_cls == "auto":
